Making paths abstract

+++ JRomManager/src-server/jrm/fullserver/handlers/UploadServlet.java
@@ -9,7 +9,6 @@ import java.nio.file.Files;
 import java.nio.file.InvalidPathException;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 
 import javax.servlet.ServletException;
@@ -20,6 +19,10 @@ import javax.servlet.http.HttpServletResponse;
 import org.apache.commons.io.IOUtils;
 
 import com.google.gson.Gson;
+
+import jrm.security.PathAbstractor;
+import jrm.server.shared.WebSession;
+import lombok.val;
 
 @SuppressWarnings("serial")
 public class UploadServlet extends HttpServlet
@@ -36,6 +39,9 @@ public class UploadServlet extends HttpServlet
 		switch(req.getRequestURI())
 		{
 			case "/upload/":
+			{
+				val ws = (WebSession)req.getSession().getAttribute("session");
+				val pathAbstractor = new PathAbstractor(ws);
 				final Result result = new Result();
 				String init = req.getParameter("init");
 				if(init != null && init.equals("1"))
@@ -44,51 +50,59 @@ public class UploadServlet extends HttpServlet
 					result.extstatus="continue...";
 					final String filename =  URLDecoder.decode(req.getHeader("x-file-name"), "UTF-8");
 					final String fileparent = URLDecoder.decode(req.getHeader("x-file-parent"), "UTF-8");
-					long filesize;
-					try
+					if(pathAbstractor.isWriteable(fileparent))
 					{
-						filesize = Long.parseLong(req.getHeader("x-file-size"));
-					}
-					catch(NumberFormatException e)
-					{
-						filesize = -1;
-					}
-					try
-					{
-						final Path dest = Paths.get(fileparent);
-						if(!(Files.exists(dest) && Files.isDirectory(dest)))
+						long filesize;
+						try
 						{
-							result.status=6;
-							result.extstatus="Error: destination " + dest + " must be an existing directory";
+							filesize = Long.parseLong(req.getHeader("x-file-size"));
 						}
-						else
+						catch(NumberFormatException e)
 						{
-							final FileStore fs = Files.getFileStore(dest);
-							final long free = fs.getUsableSpace();
-							if(free < filesize)
+							filesize = -1;
+						}
+						try
+						{
+							final Path dest = pathAbstractor.getAbsolutePath(fileparent);
+							if(!(Files.exists(dest) && Files.isDirectory(dest)))
 							{
-								result.status=7;
-								result.extstatus="Error: not enough free space, need " + filesize + " but only " + free + " is available";
+								result.status=6;
+								result.extstatus="Error: destination " + dest + " must be an existing directory";
 							}
 							else
 							{
-								final Path filepath = dest.resolve(filename);
-								Files.getLastModifiedTime(filepath);
+								final FileStore fs = Files.getFileStore(dest);
+								final long free = fs.getUsableSpace();
+								if(free < filesize)
+								{
+									result.status=7;
+									result.extstatus="Error: not enough free space, need " + filesize + " but only " + free + " is available";
+								}
+								else
+								{
+									final Path filepath = dest.resolve(filename);
+									Files.getLastModifiedTime(filepath);
+								}
 							}
 						}
+						catch(NoSuchFileException e)
+						{	// that's ok
+						}
+						catch(InvalidPathException e)
+						{	// invalid filename (case 1)
+							result.status=8;
+							result.extstatus=e.getMessage();
+						}
+						catch(FileSystemException e)
+						{	// invalid filename (case 2)
+							result.status=9;
+							result.extstatus=e.getMessage();
+						}
 					}
-					catch(NoSuchFileException e)
-					{	// that's ok
-					}
-					catch(InvalidPathException e)
-					{	// invalid filename (case 1)
-						result.status=8;
-						result.extstatus=e.getMessage();
-					}
-					catch(FileSystemException e)
-					{	// invalid filename (case 2)
-						result.status=9;
-						result.extstatus=e.getMessage();
+					else
+					{
+						result.status=11;
+						result.extstatus="Is read only";
 					}
 				}
 				else
@@ -100,6 +114,7 @@ public class UploadServlet extends HttpServlet
 				resp.setStatus(HttpServletResponse.SC_OK);
 				resp.getWriter().write(new Gson().toJson(result));
 				break;
+			}
 			default:
 				super.doPost(req, resp);
 				break;
@@ -112,48 +127,61 @@ public class UploadServlet extends HttpServlet
 		switch(req.getRequestURI())
 		{
 			case "/upload/":
+			{
+				val ws = (WebSession)req.getSession().getAttribute("session");
+				val pathAbstractor = new PathAbstractor(ws);
 				final Result result = new Result();
 				final String filename =  URLDecoder.decode(req.getHeader("x-file-name"), "UTF-8");
 				final String fileparent = URLDecoder.decode(req.getHeader("x-file-parent"), "UTF-8");
-				long filesize;
-				try
+				if(pathAbstractor.isWriteable(fileparent))
 				{
-					filesize = Long.parseLong(req.getHeader("x-file-size"));
-				}
-				catch(NumberFormatException e)
-				{
-					filesize = -1;
-					System.err.println(filename + " : " + e.getMessage());
-				}
-				Path filepath = Paths.get(fileparent, filename);
-				Files.createDirectories(filepath.getParent());
-				long size = 0;
-				try (BufferedOutputStream out = new BufferedOutputStream(Files.newOutputStream(filepath, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING, StandardOpenOption.WRITE)))
-				{
-					size = IOUtils.copy(req.getInputStream(), out);
-					result.status = 3;
-					result.extstatus = filename + " done";
-				}
-				catch(IOException e)
-				{
-					result.status = 20;
-					result.extstatus = filename + " : " + e.getMessage();
-				}
-				finally
-				{
-					if(filesize >= 0 && size != filesize)
+					long filesize;
+					try
 					{
-						result.status = 21;
-						result.extstatus = "Error: " + filename + " size should be " + filesize + " bytes long but got " + size + " bytes";
+						filesize = Long.parseLong(req.getHeader("x-file-size"));
+					}
+					catch(NumberFormatException e)
+					{
+						filesize = -1;
+						System.err.println(filename + " : " + e.getMessage());
+					}
+					final Path dest = pathAbstractor.getAbsolutePath(fileparent);
+					final Path filepath = dest.resolve(filename);
+					Files.createDirectories(filepath.getParent());
+					long size = 0;
+					try (BufferedOutputStream out = new BufferedOutputStream(Files.newOutputStream(filepath, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING, StandardOpenOption.WRITE)))
+					{
+						size = IOUtils.copy(req.getInputStream(), out);
+						result.status = 3;
+						result.extstatus = filename + " done";
+					}
+					catch(IOException e)
+					{
+						result.status = 20;
+						result.extstatus = filename + " : " + e.getMessage();
+					}
+					finally
+					{
+						if(filesize >= 0 && size != filesize)
+						{
+							result.status = 21;
+							result.extstatus = "Error: " + filename + " size should be " + filesize + " bytes long but got " + size + " bytes";
+						}
+					}
+					if(result.status != 3)
+					{
+						System.err.println(result.status + " : " + result.extstatus);
+						Files.delete(filepath);
 					}
 				}
-				if(result.status != 3)
+				else
 				{
-					System.err.println(result.status + " : " + result.extstatus);
-					Files.delete(filepath);
+					result.status=11;
+					result.extstatus="Is read only";
 				}
 				resp.getWriter().write(new Gson().toJson(result));
 				break;
+			}
 			default:
 				super.doPut(req, resp);
 				break;
